# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/exec_result":{"get":{"summary":"Get Exec Result","operationId":"get_exec_result_exec_result_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/problems/{data_type}":{"get":{"summary":"Get Problems","operationId":"get_problems_problems__data_type__get","parameters":[{"name":"data_type","in":"path","required":true,"schema":{"type":"string","title":"Data Type"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/search/{keyword}":{"get":{"summary":"Search","operationId":"search_search__keyword__get","parameters":[{"name":"keyword","in":"path","required":true,"schema":{"type":"string","title":"Keyword"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/search_site/{keyword}":{"get":{"summary":"Search Site","operationId":"search_site_search_site__keyword__get","parameters":[{"name":"keyword","in":"path","required":true,"schema":{"type":"string","title":"Keyword"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/public_feeds/{feed_name}":{"delete":{"summary":"Remove Public Feed","operationId":"remove_public_feed_public_feeds__feed_name__delete","parameters":[{"name":"feed_name","in":"path","required":true,"schema":{"type":"string","title":"Feed Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{group_name}/site_config":{"get":{"summary":"Get Site Config","operationId":"get_site_config_groups__group_name__site_config_get","parameters":[{"name":"group_name","in":"path","required":true,"schema":{"type":"string","title":"Group Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Save Site Config","operationId":"save_site_config_groups__group_name__site_config_put","parameters":[{"name":"group_name","in":"path","required":true,"schema":{"type":"string","title":"Group Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{group_name}/toggle":{"put":{"summary":"Toggle Group","operationId":"toggle_group_groups__group_name__toggle_put","parameters":[{"name":"group_name","in":"path","required":true,"schema":{"type":"string","title":"Group Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{group_name}/feeds/{feed_name}/htmls/{html_file_name}":{"delete":{"summary":"Remove Html File","operationId":"remove_html_file_groups__group_name__feeds__feed_name__htmls__html_file_name__delete","parameters":[{"name":"group_name","in":"path","required":true,"schema":{"type":"string","title":"Group Name"}},{"name":"feed_name","in":"path","required":true,"schema":{"type":"string","title":"Feed Name"}},{"name":"html_file_name","in":"path","required":true,"schema":{"type":"string","title":"Html File Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{group_name}/feeds/{feed_name}/htmls":{"delete":{"summary":"Remove Html","operationId":"remove_html_groups__group_name__feeds__feed_name__htmls_delete","parameters":[{"name":"group_name","in":"path","required":true,"schema":{"type":"string","title":"Group Name"}},{"name":"feed_name","in":"path","required":true,"schema":{"type":"string","title":"Feed Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{group_name}/feeds/{feed_name}/run":{"post":{"summary":"Run","operationId":"run_groups__group_name__feeds__feed_name__run_post","parameters":[{"name":"group_name","in":"path","required":true,"schema":{"type":"string","title":"Group Name"}},{"name":"feed_name","in":"path","required":true,"schema":{"type":"string","title":"Feed Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{group_name}/feeds/{feed_name}/toggle":{"put":{"summary":"Toggle Feed","operationId":"toggle_feed_groups__group_name__feeds__feed_name__toggle_put","parameters":[{"name":"group_name","in":"path","required":true,"schema":{"type":"string","title":"Group Name"}},{"name":"feed_name","in":"path","required":true,"schema":{"type":"string","title":"Feed Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{group_name}/feeds/{feed_name}/list":{"delete":{"summary":"Remove list","operationId":"remove_list_groups__group_name__feeds__feed_name__list_delete","parameters":[{"name":"group_name","in":"path","required":true,"schema":{"type":"string","title":"Group Name"}},{"name":"feed_name","in":"path","required":true,"schema":{"type":"string","title":"Feed Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{group_name}/feeds/{feed_name}/check_running":{"get":{"summary":"Check Running","operationId":"check_running_groups__group_name__feeds__feed_name__check_running_get","parameters":[{"name":"group_name","in":"path","required":true,"schema":{"type":"string","title":"Group Name"}},{"name":"feed_name","in":"path","required":true,"schema":{"type":"string","title":"Feed Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{group_name}/feeds/{feed_name}":{"get":{"summary":"Get Feed Info","operationId":"get_feed_info_groups__group_name__feeds__feed_name__get","parameters":[{"name":"group_name","in":"path","required":true,"schema":{"type":"string","title":"Group Name"}},{"name":"feed_name","in":"path","required":true,"schema":{"type":"string","title":"Feed Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Post Feed Info","operationId":"post_feed_info_groups__group_name__feeds__feed_name__post","parameters":[{"name":"group_name","in":"path","required":true,"schema":{"type":"string","title":"Group Name"}},{"name":"feed_name","in":"path","required":true,"schema":{"type":"string","title":"Feed Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Feed Info","operationId":"delete_feed_info_groups__group_name__feeds__feed_name__delete","parameters":[{"name":"group_name","in":"path","required":true,"schema":{"type":"string","title":"Group Name"}},{"name":"feed_name","in":"path","required":true,"schema":{"type":"string","title":"Feed Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{group_name}/feeds":{"get":{"summary":"Get Feeds By Group","operationId":"get_feeds_by_group_groups__group_name__feeds_get","parameters":[{"name":"group_name","in":"path","required":true,"schema":{"type":"string","title":"Group Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{group_name}":{"delete":{"summary":"Remove Group","operationId":"remove_group_groups__group_name__delete","parameters":[{"name":"group_name","in":"path","required":true,"schema":{"type":"string","title":"Group Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups":{"get":{"summary":"Get Groups","operationId":"get_groups_groups_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}